added driver files for crc and flash
